Added example usage and formatted long lines in README
The README did not have clear enough examples to really show off the
awesomeness of this project. So I added a few examples that will quickly
show how useful this tool is.

Also there were just a few very long lines in the README so I formatted
them to around 80 characters.