Allow output paths to contain dots
Change the regex to only find a '.' if it is followed by md or svg, so
as to be confident that the dot was part of the extension and not
earlier in the path.

The way the regex was written to increment the -n part of the SVG
filename it would incorrectly put said number after the first dot
encountered in the output path and crash the program.