build: switch package.json to type "module"
This declares that this package is an ESM module,
and that all .js files should be expected to be ESM.

Because src/index.js is already valid ESM code for Node.JS 14,
I've removed babel, since it's not needed.

BREAKING CHANGE: This module is ESM only.
BREAKING CHANGE: The input to this module is now in src/index.js
BREAKING CHANGE: The `prepublishOnly` script has been removed,
                 as it is no longer needed without babel.