fix: add css as <style> node to SVG
If a --cssFile is given to mermaid-cli, adds it as a `<svg><style>`
node. This means it's part of the `<svg>` and will get exported
correctly when creating a SVG file.

See https://developer.mozilla.org/en-US/docs/Web/API/SVGStyleElement
for documentation on the WebAPI on the `<svg><style>` element.

Co-authored-by: Joel Burton <joel@joelburton.com>
Co-authored-by: Niclas Gustafsson <niclas@bitfront.se>
Co-authored-by: Mindaugas Laganeckas <Mindaugas.Laganeckas@3shape.com>