fix: fix page.viewport() race-condition
For some reason, the tests in GitHub Actions are failing about 1/2
the time. I'm not 100% sure what the issue is, but I think it's a
race condition, and page.viewport() is supposed to be `await`-ed,
so maybe it's that?

I've also re-arannged the test-code in case it's that too.