Revert "Access to secrets in Prs from forks"
This reverts commit 0e2c06db0dbe36e269bb1fd48ebbd3cc5b05234a.

The `pull_request_target` event does have access to the repository
secrets, but this is because it doesn't run on the fork contents.

Instead, it runs on the target of the PR,
e.g. when somebody makes a PR, instead of testing/running on their
fork's `feature/...` branch, it runs on `mermaid-cli`'s `master`
branch, which hasn't been changed.

Unfortunately, there's no easy way to give forks access to our
GitHub Actions secrets, because if there was, they could just do
`"echo "$MY_SECRET"` and they'll see the secret.