docs: document running Percy on PRs from forks
Essentially, in order to get percy running on PRs from forks,
we can just fetch the PR branch from the fork, then push it to a new
branch on the main mermaid-cli repo.

Since Percy uploads happen on `push` too, GitHub should automatically
show the percy results on the PR, as long as both the fork PR and
the mermaid-cli branch have the exact same commit hash.