docs: document `--cssFile` option in README
Adds an example of how to use the `--cssFile` option.

Additionally, I've added an example output to the docs folder,
which will get automatically regenerated whenever `yarn test` is run.
(it uses the [`deterministicIds` mermaid options][1], so it should
 very rarely change (except when mermaid gets updated))

This output is hidden by a `<details></details>` HTML block in the
README.md, which is valid [GitHub Flavored Markdown][2], in order to
avoid taking up too much space, and blinding people if they're
using Dark Mode on GitHub.

[1]: https://github.com/mermaid-js/mermaid/blob/develop/docs/Setup.md#deterministicids
[2]: https://github.github.com/gfm/#html-block