test: make animated SVG static for first 3seconds
Currently, the flowchart1-with-css.svg/.png file is constantly
animated. Since there is a slight bit of random delay when converting
this animated SVG to a static picture, each PNG file is slightly
different, causing Percy to throw an error, showing that the files
have changed slightly.

This commit makes the animated SVG static for the first 3 seconds,
allowing for mermaid-cli and convert-svg-to-png to always take the
same snapshot.