feat: load title/desc text from Mermaid SVG
Loads the title and/or description text from the Mermaid SVG for
accessibility purposes.

In come countries, it is illegal to use images on a website without
the appropriate alt text. Additionally, some markdown renderers
use the alt text as a caption (e.g. Pandoc).

For example, the following mermaid code:

```mermaid
stateDiagram
  accTitle: State diagram title here
  accDescr: State diagram description here

  state Choose <<fork>>
  Choose --> Still
```

Will get converted to the following markdown image:

```markdown
![State diagram description here](./link-to-image.png "State diagram title here")
```

[] in the alt text and " in the title text are automatically escaped.