ci(test): fix convert-svg-to-png in mermaid 9.1.7
Mermaid v9.1.7 implemented commit
https://github.com/mermaid-js/mermaid/commit/2968b400c49f66983e6d206959504e72aa6ae0ef
that removes the `width` attr from SVGs by default,
using the CSS max-width style instead.

Percy CI (used for visual regression tests) does not yet support SVGs,
so we use convert-svg-to-png@v0.5.0, but that requires either the width
or the height to calculate the size of the PNG to convert to SVG to.

To fix this, the Mermaid useMaxWidth option can be set to false.
Annoyingly, we need to do this for every single graph type,
see https://mermaid-js.github.io/mermaid/#/Setup?id=usemaxwidth