fix: improve stdin input handling
Improve how mmdc handles stdin inputs.

Currently, running `mmdc` exits with an error if it cannot find
an stdin. However, it does this by checking if the stdin is a FIFO,
which is only true in some cases.
(it seems to fail on Windows and other non-bash shells).

I've changed `mmdc` to always read from stdin if there is no `--input`.
However, it prints a warning.

To suppress this warning, you can run mmdc with `--input -`.