test: update flowchart3 test case
Mermaid v10.2.0 changed some behavior for htmlLabels with embedded
SVG images. The `width=100%` is now set in the node CSS, which normally
scales the images to the width of the text, but for this SVG, it seems
to crop it. See https://github.com/mermaid-js/mermaid/issues/4455.

Setting a well-defined `viewBox` attribute to our SVG fixes this.

Co-authored-by: Laura Valentine Tscharner <laura.tscharner@tecnico.ulisboa.pt>