Hide "Puppeteer old Headless deprecation warning"
Hide the "Puppeteer old Headless deprecation warning" that is printed
in Puppeteer v19.11.0 or later, when `headless: "new"` is not used.

Using `headless: "new"` currently isn't very stable in Puppeteer
v19.11.0, and we can't upgrade to Puppeteer v20 until mermaid-cli drops
Node.JS v14 support.

Instead, we can set `headless: 1` as the default value. This is not
officially supported, but puppeteer only prints the warning if the
[`headless` option is `true`][1], but it still uses headless mode
if the [`headless` option is not `"new"` and truthy][2].

[1]: https://github.com/puppeteer/puppeteer/blob/159513c8dbe2c9f51aa37dbe531d52b5daf1e106/packages/puppeteer-core/src/node/ChromeLauncher.ts#L50
[2]: https://github.com/puppeteer/puppeteer/blob/159513c8dbe2c9f51aa37dbe531d52b5daf1e106/packages/puppeteer-core/src/node/ChromeLauncher.ts#L213-L215