Fix link to fork "button"
A /fork leads to link on the GitHub webview of the markdown,
which ends with mermaid-cli/blob/master/fork. This is the link 
to a directory called "fork" in root of the repository.
So go back to parent directory, two times, to get mermaid-cli/fork,
which is the link to the fork "button" of the current repository.