build: remove single-file inlining
Compile JavaScript code as an IIFE, allowing us to avoid needing to
inline anything.

This makes the code about 10% faster for me
(tests run in 105s instead of 120s), and the size on the bundle went
from 4MiB to 3.5MiB.

Additionally, I think I've finally figured out the issue that was
sometimes causing the FontAwesome icons to load incorrectly: sometimes
Puppeteer was parsing the CSS FontAwesome as Latin-1 instead of UTF-8!
Explicility setting the script to have `charset="utf-8"` seems to fix
this!