Support using stdout as output
If the user specifies `-` as the output, the code will be written to
`stdout` instead of writing to a file.

Note that for the sake of making the code more readable, I'm using
explicit `'/dev/stdout'`.  This is just a personal preference and
doesn't have anything to do with platform.