ci: fix skipping **Upload to Percy** on errors
Right now, when a test fails, the **Upload to Percy** CI job still runs,
and fails.

I think this is because our syntax for the `if:` statement was wrong and
had an extra newline/spaces, that was making GitHub Actions treat the
statement as always `true`.

According to
https://github.com/orgs/community/discussions/25641#discussioncomment-10617902,
apparently `>-` and removing the outer `${{` and `}}` might help.